Agregue titulo y descripcion. función update para actualizar los datos de perfil

+++ src/main/java/com/bprofilefrangimz/bprofile/Interface/IPersonaService.java
@@ -11,8 +11,8 @@ public interface IPersonaService {
     public void savePersona(Persona persona);
 
     //Eliminar un objeto buscandolo por id
-    public void deletePersona(Long id);
+    public void deletePersona(int id);
 
     //Buscar una persona por id
-    public Persona findPersona(Long id);
+    public Persona findPersona(int id);
 }
